update learner for data loading
